refactor: rename chats package to home
