+++ core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/di/viewModelModule.kt
@@ -3,8 +3,8 @@
 package net.thechance.mena.core_chat.presentation.di
 
 import net.thechance.mena.core_chat.presentation.screen.chat.ChatViewModel
-import net.thechance.mena.core_chat.presentation.screen.chats.HomeViewModel
 import net.thechance.mena.core_chat.presentation.screen.contacts.ContactsViewModel
+import net.thechance.mena.core_chat.presentation.screen.home.HomeViewModel
 import net.thechance.mena.core_chat.presentation.screen.syncContacts.SyncContactsViewModel
 import org.koin.core.module.dsl.viewModel
 import org.koin.core.qualifier.named
